refactor: user 패키지 분리
