+++ src/main/kotlin/lotto/model/GameException.kt
@@ -1,5 +1,7 @@
 package lotto.model
 
+import lotto.model.user.UserEvent
+
 class GameException(
     message: String,
     val event: UserEvent
